% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{addTitle}
\alias{addTitle}
\title{label plots}
\usage{
addTitle(
  object,
  text,
  color = "black",
  fontSize = "20px",
  fontFamily = "Sans",
  leftPosition = 50,
  topPosition = 2
)
}
\arguments{
\item{object}{leaflet plot}

\item{text}{text}

\item{color}{text color}

\item{fontSize}{font size}

\item{fontFamily}{font}

\item{leftPosition}{left position}

\item{topPosition}{top position}
}
\value{
title on plots
}
\description{
label plots
}
