% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_clus_kml.R
\name{ind_clus_kml}
\alias{ind_clus_kml}
\title{Plot individual cluster .kml}
\usage{
ind_clus_kml(
  AID,
  cn,
  locs,
  cs,
  centroid_calc = "mean",
  overwrite = TRUE,
  dir = NULL
)
}
\arguments{
\item{AID}{Desired AID from sequential cluster output}

\item{cn}{Desired cluster number}

\item{locs}{Location dataframe output from GPSeq_clus()}

\item{cs}{Cluster summary output from GPSeq_clus()}

\item{centroid_calc}{'mean' (default) or 'median' centroid plot}

\item{overwrite}{TRUE (default) labels output as "ind.kml" that overwrites with each run within tempdir(). FALSE saves outputs as "AID_cn"}

\item{dir}{File path when saving output}
}
\value{
Opens the cluster locations and centroid .kml for assessment.
}
\description{
Uses results from 'GPSeq_clus" to plot individual cluster .kmls
}
