% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{julian_conv}
\alias{julian_conv}
\title{Julian Conversion}
\usage{
julian_conv(x)
}
\arguments{
\item{x}{vector of input dates}
}
\value{
vector of julian days
}
\description{
Julian Conversion
}
