% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{ML_ex_dat}
\alias{ML_ex_dat}
\title{Sample Data for Sequential Clustering Routine}
\format{
A \code{dataframe} containing 4 columns:
\describe{
\item{AID}{animal identification}
\item{TelemDate}{location timestamp in POSIXct format}
\item{Lat}{latitude coordinates}
\item{Long}{longitude coordinates}
}
}
\usage{
ML_ex_dat
}
\description{
A \code{dataframe} containing a subset of GPS location data from 2 male and 1 female mountain lions used
for testing and running sequential cluster function examples.Example data provided by Wyoming Game and Fish Department, 2020.
}
\keyword{datasets}
