% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{moveMe}
\alias{moveMe}
\title{arrange columns}
\usage{
moveMe(data, tomove, where = "last", ba = NULL)
}
\arguments{
\item{data}{input dataframe}

\item{tomove}{which column(s) to move}

\item{where}{where to move them - e.g. "before", "after", "first", "last"}

\item{ba}{??}
}
\value{
Dataframe with new column order
}
\description{
arrange columns
}
